move api code into api app
