+++ message_coding/apps/base/models.py
@@ -1,12 +1,5 @@
 from django.db import models
 
-# Use get_user_model() to get the user model
-# i.e. do not import django.contrib.auth.models.User directly!
-# You can also use settings.AUTH_USER_MODEL, in ForeignKeys for example.
-from django.contrib.auth import get_user_model
-from django.conf import settings
-
-
 class NameDescriptionMixin(models.Model):
     """Inherit from this to add name/description fields"""
